(tests) Testsuite: avoid API query in loggedInAs()
This API query was only needed for RealHttpEngine (which is currently
not working), and is not needed for CliEngine (which we use now).

Therefore abstract ModerationTestsuiteEngine would only keep track of
$currentUser and assume that its subclass double-checks the rest.

Also $currentUser was moved away from ModerationTestsuite class as
unnecessary duplicate (it's sufficient to have it in the Engine class).