(tests) Use ModifyDbRowTestTrait in ActionsHaveConsequencesTest
Here we provide makeDbRow($fields) method for use in Consequence tests.

There is no point to call doEditContent() and other high-level methods
in ActionsHaveConsequencesTest::setUp(), because the fact that a DB row
with correct fields is created after doEditContent() is already covered
by other tests:
- EditsHaveConsequencesTest
- QueueEditConsequenceTest
- InsertRowIntoModerationTableConsequenceTest

Instead we should just insert a row directly, same as
in [30_action/ModerationActionTest.php] of the blackbox testsuite.