(tests) Fix test failure in 1.31 caused by dirty state before the test
It was noticed that in MediaWiki 1.31 "moderation" table had 8 rows
(instead of expected 0) at the very beginning of testApproveAll()
- that is, before any of our own code was even executed.

This means that MediaWikiTestCase class doesn't properly clean the
tables between tests, even though they are listed in $tablesUsed.

This is a problem in MediaWiki core, and it doesn't exist in MW 1.32+,
so a workaround is to just clear the tables from $tablesUsed ourselves.