(phan) Fix new Phan warnings after updating phan-config to 0.10.1
1) Phan doesn't like self::singleton()->something syntax,
so we replace this with $a = self::singleton(); $a->something.

2) Warnings in tests/phpunit/framework/ are false positives that are
not applicable to testsuite. (these checks are for production code)

For example, CliEngine prints data from external script to stdout,
but this script is our own [cliInvoke.php], so this data is not really
external. Additionally, this is testsuite, and stdout is actually what
it is, stdout. This code is not used on web pages or anything.