(tests) .travis.yml: trusty -> xenial
This fixes "git clone" being unable to get extensions from Gerrit.
Version of gnutls in Ubuntu 14 (Trusty) uses obsolete/insecure TLS 1.1,
which is no longer supported by https://gerrit.wikimedia.org/

Note: memcached was started by default in trusty, but not in xenial.
So we add it to "services to start" explicitly.