(cosmetic) Don't use wrapWikiMsg() when $wgTitle can be undefined
If some third-party maintenance script calls EditFormPreloadText hook
(in the same way as ApiQueryInfo does), then global Title may be null.

Method OutputPage::wrapWikiMsg() is implemented in such a way that it
requires global Title to be set, and throws an exception otherwise.

Therefore we shouldn't use wrapWikiMsg() in showPendingEdit().