(bug) Fix error when editing Flow forums in MW 1.32+
This affected two situations:
1) edited page is not TextContent.
2) ModerationIntercept handler returns false ("don't intercept this").

Cause:
1) in ModerationEditHooks::onPageContentSave() we converted $comment
from CommentStoreComment to string.
2) if after that we decide to not intercept the edit (e.g. because
the Content is not TextContent - for example, a Flow forum),
then $comment will remain "string", because onPageContentSave()
received $comment parameter it by reference.
3) PageUpdater::saveRevision() would use this $comment (string) with
the assumption that it is a CommentStoreComment. (this will fail)

Solution is to not receive $comment parameter by reference,
as we don't really need or want to modify this variable in the caller.

While we are at it, we also:
1) remove reference from all other parameters (we don't need to modify
any of them either - with the exception of $status, which is an object
and will still be modified by $status->fatal()).
2) replace is_a() with "instanceof".