(lib) Make ApproveHook work with moves with suppressed redirects
Note: currently approval of moves always creates a redirect, so this
doesn't change anything in normal behavior of Moderation.

However, it was noticed that suppressing a redirect made ApproveHook
not be applied to the move (because scheduleDoUpdate() wasn't called
from modifyRedirectAfterMove() due to the absence of said redirect).

We are fixing this for consistency.