(cleanup) Remove unused code in ModerationActionShowImage class
The following unneeded code was discovered and removed:

1) selecting mod_user and mod_user_text is no longer necessary.
Previously these fields were needed to find uploader's UploadStash,
but we no longer use it (instead we use the stash of system user).

2) Condition of removed "if ( !$file )" always evaluates to false,
because UploadStash::getFile() always returns a valid File and throws
an exception if this can't be done. Additionally in the code below
variable $file can be set to "new UnregisteredLocalFile",
but this will not be cast to false either.