(refactor) Remove InstallApproveHookConsequence - no longer needed
This consequence class was previously needed for testsuite to ensure
that ApproveHook has in fact been installed with expected parameters.

It is no longer needed, because:
1) ApproveHook is dependency-injected into ModerationApprovableEntry,
2) we have proper unit tests (and can replace ApproveHook with a mock),
which allows us to detect the call to ModerationApproveHook::addTask()
and check its parameters.
3) InstallApproveHookConsequence does nothing else except addTask().
4) ApproveHook is now a resettable service, so calling addTask() in one
of the tests won't have any effect on the following tests.