(compat) Compatibility with MySQL 5.7: don't use DEFAULT on BLOB field
See issue #56
("Installation Error: 'mod_tags' can't have a default value").

MySQL 5.7 doesn't support DEFAULT for BLOB fields, such as mod_tags.
This is not a problem (DEFAULT clause can be safely omitted), because
the default value without it is NULL, which is also correct/acceptable.

Note: MariaDB is unaffected (it does support DEFAULT for BLOB fields).

Note: we are not adding a patch for existing setups, because they are
already working (due to using MariaDB?) and because presence or absence
of this DEFAULT is irrelevant (see above).