(compat) Workaround for T258159 (MW web updater doesn't load services)
See issue #56.
It was noticed that getService('Moderation.VersionCheck') fails when
using the MediaWiki web updater, because the web updater doesn't load
extension-provided services (unlike when running update.php directly).

This is a bug in MediaWiki core.

A workaround is to check "is the service loaded", and if not, then load
our ServiceWiring.php directly.