(tests) Replace assertContains() with assertStringContainsString()
Newer PHPUnit used in MediaWiki 1.35+ prints an error if assertContains
is used on two strings (it can only be used on element and array).

Thanks to our polyfill class CompatPHPUnit6Trait, we can (and should)
use this method even with the older versions of PHPUnit.
(which are used by MediaWiki 1.31-1.34)