(tests) Don't use --reuse-db (broken in MediaWiki 1.36, see T283949)
In MediaWiki core (1.36), MediaWikiIntegrationTestCase class is now
always deleting the tables that were created for tests,
even in --use-normal-tables mode (that is supposed to preserve them).

Luckily it doesn't break CliEngine (it uses those tables before they
are deleted).

However, it completely breaks the approach where we first call a small
test (to create the tables), and then quickly run the remaining tests
with --reuse-db without the need to recreate tables for every test.