(compat) Don't intercept edits in namespace of Extension:CommentStreams
In Extension:CommentStreams, CommentStreamsStore::updateComment()
calls doEditContent() and then misinterprets the return code "edit was
sent to moderation" as an error (instead of "saved successfully"),
and then it panics "we couldn't save the comment".

Because this breaks CommentStreams, a workaround on Moderation side
is to not intercept edits in CommentStreams namespace at all
(same approach as with Flow forums).