(1.37) ApproveHook: fix calculation of rev_id of redirects in 1.37
This fixes failure of testMoveAndEditWithSameUserAndPage() in 1.37,
which covers situation when the same user A edited and renamed the same
page B, which was approved simultaneously with "Approve all".

Investigation shows that in MediaWiki 1.37 the PageMoveComplete hook
was receiving a wrong revision ID of the (newly created) redirect page
(only during the automated testsuite, not in manual tests).
Instead of rev_id of redirect (3), we got rev_id of the null edit (4).

It looks as if null edit (4) is marked as "next revision" for (3)
some time after the renaming.

Solution is to replace PageMoveComplete (which is deferred and happens
after these "next revision" modifications) with PageMoveCompleting,
which runs immediately upon the successful move (is not deferred).