(1.38) Update ApproveHook for MediaWiki 1.38
This commit reimplements ApproveHook to not depend on the order
of DeferredUpdates. Previous order (in MediaWiki 1.35-1.37) allowed us
to call doUpdate() after ALL of RecentChange_save hook invocations.
This was efficient (less SQL queries), but it wouldn't work in 1.38.

Note that MediaWiki 1.35-1.37 still use the previous implementation,
so this change doesn't result in any performance impact for 1.35-1.37.