(visualeditor) Fix VE in 1.40 to allow preloading of new page creations
Before this fix, VisualEditor in MW 1.40 was displaying an error when
the user tried to edit a not-yet-approved page that was just created
by this user. Edits in existing pages didn't have this problem.

This is an imperfect fix: we are counting on VisualEditor's new feature
(VE now saves the last edit in the local storage) to preload for us.

Ideally, Moderation should be preloading this,
but VisualEditor keeps adding more and more sanity checks that prevent
us from doing so on pages that don't yet exist.