(tests) Provide CSRF token to blackbox tests of ApiBeforeMain hook
Before MediaWiki 1.41, this wasn't needed, because CliEngine hook
from InvokedWikiSettings.php was called before the hook that we are
testing (ModerationApiHooks::onApiBeforeMain).

Because of how wfFakeHooksRegister() works, the order of these 2 hooks
was reversed, and now DerivativeRequest is created before CliEngine
has a chance to replace the CSRF token placeholder.

Therefore we replace the placeholder for the second time, in the test's
ApiBeforeMain handler, which has access to this DerivativeRequest.