(1.41) Fix deprecation warnings ("targets" parameter of ResourceLoader)
Parameter "targets" is deprecated (all modules should be both desktop
and mobile), and will be removed in MediaWiki 1.42.

We still don't need scripts like [notify.mf.css] on desktop (because
it's a script specifically for MobileFrontend interface),
but the checks before addModules() allow us to control this anyway.

Not covering this with the test (unfortunately), because MobileFrontend
uses and modifies too many global variables to be invoked in unit tests
(not without a ton of code to meticulously reset all these changes).

Resolves #84.