Truncate extremely long edit summaries (>255 bytes)
Normal editing form already does this, but if the user somehow submits
a longer summary, it is now silently truncated, and edit is saved
(instead of showing scary "database error" to the user).

Previously this problem didn't exist, because MediaWiki core itself
had a limit of 255 bytes, but turns out, $summary has NOT yet been
truncated at the moment when MultiContentSave hook is called.