(tests) Correctly parse HTML pages that are preceded by PHP warnings
Previously the blackbox tests would fail to parse HTML document in the
response if it was preceded by PHP deprecation warnings, etc.

Now the deprecation warning will be printed separately (still causing
the test to be considered "Risky" rather than successful),
but the parsing of HTML document won't be affected.