(1.43) Adjust position of postedit notification in Vector 2022
In MediaWiki 1.43 (which has temporary accounts enabled by default),
the header "You are logged in as <temporary account>" can cover a part
of our notification "Your edit was queued for moderation".

This notification is essential for new users.

Because the notification can be closed (on click) and header can't be,
the notification should be on top.