(tests) Remove RealHttpEngine (hasn't been used since MediaWiki 1.28)
Before MediaWiki 1.28 started sandboxing the tests, RealHttpEngine was
fully functional and, in many ways, preferable (because it's closest to
how real users are accessing the wiki, and also the simplest).

MediaWiki 1.28 made RealHttpEngine completely unusable. Numerous
workarounds were attempted to fix it, but the following versions
of MediaWiki kept isolating the tests more and more,
to the point where RealHttpEngine (even if we could fix it) wouldn't be
any simpler than CliEngine.

With CliEngine being fully functional (successfully used in 1.28-1.43),
there is no point in keeping non-working RealHttpEngine in the codebase.