(feature) Auto-reject null edits when moderator clicks Approve on them
"Null edit" means that this pending change is pointless to Approve,
because the page already has the exact same text
as what this pending change was proposing to add.

Because Approve is impossible, moderator had no choice but to click
Reject on this change to move it from Pending to Rejected folder.

Now this is done automatically, moderator doesn't need to click.