MarkAsMergedConsequence should ignore changes without edit conflict
This is a minor sanity check.

It prevents changes that don't have an edit conflict
(can be Approved directly, without having to use Merge)
from being marked as "merged".

This was never happening in normal use of the extension anyway
(action=merge prints an error if mod_conflict=0),
unless the moderator manually adds wpMergeID to the edit form.