(mobile) Fix some postedit notifications having a link to the Main Page
Depending on how the edit form is visited, MobileFrontend is sometimes
redirecting the user to the wrong URL: /w/index.php without title, which
points to the Main Page instead of the article that was just edited.

In this case, we can't rely on current URL when forming "your version of
this page" link in postedit notification.
Instead we need to determine the title from the Referrer.